BackEnd/requirements.txt
asgiref==3.8.1
boto3==1.35.78
botocore==1.35.78
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.10
jmespath==1.0.1
psycopg==3.2.3
psycopg-binary==3.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
